requirements.txt
fastapi
uvicorn[standard]
httpx
pypdf
fpdf2
Jinja2
python-multipart
pywebview
